- Add MustNotUseMethodSpecification
- Reimplement MustNotUsePropertyGetterSpecification in terms of MustNotUseMethodSpecification
